Worked on value extractor
It now uses an absolute path to get the trec_eval program. You’ll
notice I suffixed it with macosx just in case there is a problem with
windows, we can add a windows complied trec_eval here and sort that
out. In the final program we can just import this file and use the
function to get the results in a dictionary.